Rework scheduler functions
A new set of functions is added: _schd_suspend_task and
_schd_resume_task that are used for suspending and resuming given task.
The task block functions: _schd_block_task and _schd_unblock_task are
now used as wrppper that work as an interface for these new functions.
Thus suspend and resume functions are not called directly but trough
these block/unblock functions.