Update context switch related functions
Before calling _schd_pend_context_switch check if the ready queue
contains at least one task. If this is not checked and queue happens to
be empty the scheduler will try to schedule non-existing task whcih will
lead system to crash. If the ready queue is empty do not pend the PendSV exception.