Add support for priority based scheduling
To support priority scheduling, task control block structure is updated with new
priority field, which is assigned during task's creation. Instead of one ready queue
now there are multple queues that are sorted by the aforementioned scheduling priority
of a task. A new priority bitmap variable is used to keep track of empty and non-empty
ready queues, which is used to schedule always the highest priority task
at that time.